requirements.txt
python-dotenv>=0.19.0
DeeperSeek